Refactor sync logic into its own module.
